refactor nano utils (3/3)
