+++ python/nano/src/bigdl/nano/deps/neural_compressor/pytorch/quantization.py
@@ -22,7 +22,7 @@ from .metric import PytorchINCMetric
 from .quantized_model import PytorchQuantizedModel
 from torchmetrics import Metric
 import torch
-from bigdl.nano.pytorch.utils import patch_attrs_from_model_to_object
+from bigdl.nano.utils.pytorch import patch_attrs_from_model_to_object
 
 
 class PytorchQuantization(BaseQuantization):
