+++ python/nano/src/bigdl/nano/deps/onnxruntime/pytorch/pytorch_onnxruntime_model.py
@@ -23,7 +23,7 @@ from bigdl.nano.pytorch.model import AcceleratedLightningModule
 from bigdl.nano.utils.pytorch import export_to_onnx
 from bigdl.nano.utils.common import invalidInputError
 from bigdl.nano.pytorch.context_manager import generate_context_manager
-from bigdl.nano.pytorch.utils import patch_attrs_from_model_to_object
+from bigdl.nano.utils.pytorch import patch_attrs_from_model_to_object
 
 
 class PytorchONNXRuntimeModel(ONNXRuntimeModel, AcceleratedLightningModule):
