+++ python/nano/src/bigdl/nano/deps/openvino/pytorch/model.py
@@ -26,7 +26,7 @@ from bigdl.nano.utils.common import invalidInputError
 from ..core.utils import save
 from torch.utils.data.dataloader import DataLoader
 from bigdl.nano.pytorch.context_manager import generate_context_manager
-from bigdl.nano.pytorch.utils import patch_attrs_from_model_to_object
+from bigdl.nano.utils.pytorch import patch_attrs_from_model_to_object
 
 
 class PytorchOpenVINOModel(AcceleratedLightningModule):
