+++ python/nano/src/bigdl/nano/pytorch/vision/transforms/transforms.py
@@ -28,7 +28,7 @@ import torchvision.transforms as tv_t
 from torchvision.transforms.functional import InterpolationMode
 import opencv_transforms.transforms as cv_t
 from bigdl.nano.utils.common import invalidInputError
-from bigdl.nano.pytorch.utils import TORCHVISION_VERSION_LESS_1_14
+from bigdl.nano.utils.pytorch import TORCHVISION_VERSION_LESS_1_14
 
 
 __all__ = [
