+++ python/nano/src/bigdl/nano/utils/pytorch/attributes.py
@@ -14,11 +14,19 @@
 # limitations under the License.
 #
 
-from bigdl.nano.utils.pytorch import TORCHVISION_VERSION_LESS_1_12
-from torchvision.datasets import *
-del ImageFolder
-if not TORCHVISION_VERSION_LESS_1_12:
-    del OxfordIIITPet
-    from .oxfordpet_datasets import OxfordIIITPet
 
-from .datasets import ImageFolder, SegmentationImageFolder
+import torch
+from torch import nn
+
+
+def patch_attrs_from_model_to_object(model: nn.Module, instance):
+    """
+    Patch non nn.Module public attributes of original nn.Module to a new instance.
+
+    :param model: a torch.nn.Module
+    :param instance: a instance of any object
+    """
+    for attr in dir(model):
+        if attr not in dir(instance) and not attr.startswith('_') and not\
+                isinstance(getattr(model, attr), torch.nn.Module):
+            setattr(instance, attr, getattr(model, attr))
