+++ python/nano/test/pytorch/tests/inference/test_ipex_jit_inference.py
@@ -26,7 +26,7 @@ from torch import nn
 import operator
 from bigdl.nano.pytorch import InferenceOptimizer
 from bigdl.nano.pytorch.vision.models import vision
-from bigdl.nano.pytorch.utils import TORCH_VERSION_LESS_1_10
+from bigdl.nano.utils.pytorch import TORCH_VERSION_LESS_1_10
 from bigdl.nano.utils.common import compare_version
 import tempfile
 from typing import List
