+++ python/nano/test/pytorch/tests/test_dispatcher.py
@@ -21,7 +21,7 @@ import pytest
 import torch
 
 from bigdl.nano.utils.common import invalidOperationError
-from bigdl.nano.pytorch.utils import TORCH_VERSION_LESS_1_10
+from bigdl.nano.utils.pytorch import TORCH_VERSION_LESS_1_10
 
 
 class TestDispatcherPytorch(TestCase):
