+++ python/nano/test/pytorch/tests/train/trainer/test_trainer_ipex.py
@@ -27,7 +27,7 @@ from torch import nn
 
 from bigdl.nano.pytorch import Trainer
 from bigdl.nano.pytorch.vision.models import vision
-from bigdl.nano.pytorch.utils import TORCH_VERSION_LESS_1_10
+from bigdl.nano.utils.pytorch import TORCH_VERSION_LESS_1_10
 
 batch_size = 256
 max_epochs = 2
