+++ python/nano/test/pytorch/utils/_train_ipex_callback.py
@@ -23,7 +23,7 @@ from pytorch_lightning.callbacks import Callback
 from pytorch_lightning.plugins.training_type import SingleDevicePlugin, DDPSpawnPlugin
 from pytorch_lightning.accelerators.cpu import CPUAccelerator
 
-from bigdl.nano.pytorch.utils import TORCH_VERSION_LESS_1_10
+from bigdl.nano.utils.pytorch import TORCH_VERSION_LESS_1_10
 from bigdl.nano.utils.common import _avx512_checker
 from bigdl.nano.utils.common import compare_version
