+++ python/nano/tutorial/training/pytorch/run_nano_pytorch_test.sh
@@ -2,7 +2,7 @@ export ANALYTICS_ZOO_ROOT=${ANALYTICS_ZOO_ROOT}
 export NANO_HOME=${ANALYTICS_ZOO_ROOT}/python/nano/src
 export NANO_TUTORIAL_TEST_DIR=${ANALYTICS_ZOO_ROOT}/python/nano/tutorial/training/pytorch
 
-TORCH_VERSION=`python -c "from bigdl.nano.pytorch.utils import TORCH_VERSION_LESS_1_12; print(TORCH_VERSION_LESS_1_12)"`
+TORCH_VERSION=`python -c "from bigdl.nano.utils.pytorch import TORCH_VERSION_LESS_1_12; print(TORCH_VERSION_LESS_1_12)"`
 
 set -e
